apps/kzen/requirements-dev.txt
absl-py==0.13.0
ansimarkup==1.5.0
appdirs==1.4.4
appengine-python-standard==0.2.4
argon2-cffi==20.1.0
astroid==2.5.6
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
bleach==3.3.0
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudevents==1.2.0
colorama==0.4.4
colorlog==5.0.1
cryptography==3.4.7
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dfcx-scrapi==1.0.7
docker==5.0.0
docopt==0.6.2
english==2020.7.0
entrypoints==0.3
executing==0.6.0
fast-json==0.3.2
Flask==2.0.1
flatbuffers==2.0
frozendict==2.0.7
functions-framework==2.2.1
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-bigquery==2.20.0
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.7.0
google-cloud-dialogflow-cx==0.7.1
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
gspread==3.7.0
gspread-dataframe==3.2.1
gspread-pandas==2.3.0
gunicorn==20.1.0
h5py==3.3.0
httplib2==0.19.1
icecream==2.1.1
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.9.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.6.0
libcst==0.3.19
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
monotonic==1.6
murmurhash==1.0.5
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
neo4j==4.3.1
neotime==1.7.4
nest-asyncio==1.5.1
networkx==2.5.1
ninja==1.10.0.post2
notebook==6.4.0
numpy==1.21.0
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandas-gbq==0.15.0
pandas-stubs==1.1.0.11
pandocfilters==1.4.3
pansi==2020.7.3
parso==0.8.2
pathspec==0.8.1
pathy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipdeptree==2.0.0
pkg_resources==0.0.0
pluggy==0.13.1
preshed==3.0.5
prettylog==0.3.0
profilehooks==1.12.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.18.1
protobuf==3.19.1
ptyprocess==0.7.0
py==1.10.0
py2neo==2021.1.5
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pydata-google-auth==1.2.0
Pygments==2.9.0
pylint==2.8.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-watch==4.2.0
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.19.0
rsa==4.7.2
ruamel.yaml==0.15.100
semantic-version==2.8.5
Send2Trash==1.7.1
six==1.16.0
smart-open==3.0.0
smmap==4.0.0
srsly==2.4.1
tabulate==0.8.9
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==8.0.6
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.5
wasabi==0.8.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xtermcolor==1.3
yapf==0.31.0
zipp==3.4.1
